salon_manager/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Authlib==1.5.1
bandit==1.8.3
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==44.0.2
distlib==0.3.9
Django==5.1.6
django-crispy-forms==2.3
django-formtools==2.5.1
django-phonenumber-field==8.0.0
dparse==0.6.4
filelock==3.16.1
identify==2.6.7
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
nltk==3.9.1
nodeenv==1.9.1
packaging==24.2
pbr==6.1.1
phonenumberslite==8.13.55
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
psutil==6.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruff==0.9.9
safety==3.3.1
safety-schemas==0.0.11
setuptools==75.8.2
shellingham==1.5.4
sqlparse==0.5.3
stevedore==5.4.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
